Refactor the text presence assessment
