+++ packages/yoastseo/spec/scoring/assessments/readability/TextPresenceAssessmentSpec.js
@@ -1,4 +1,4 @@
-import textPresence from "../../../../src/scoring/assessments/readability/textPresenceAssessment.js";
+import TextPresence from "../../../../src/scoring/assessments/readability/TextPresenceAssessment.js";
 import Paper from "../../../../src/values/Paper.js";
 import Factory from "../../../specHelpers/factory.js";
 const i18n = Factory.buildJed();
@@ -6,7 +6,7 @@ const i18n = Factory.buildJed();
 describe( "Assesses presence of text", function() {
 	it( "returns a score of 3 and a feedback string for a text under 50 words", function() {
 		const mockPaper = new Paper( "Hallo" );
-		const assessment = textPresence.getResult( mockPaper, Factory.buildMockResearcher( 0 ), i18n );
+		const assessment = new TextPresence().getResult( mockPaper, Factory.buildMockResearcher( 0 ), i18n );
 
 		expect( assessment.getScore() ).toEqual( 3 );
 		expect( assessment.getText() ).toEqual( "<a href='https://yoa.st/35h' target='_blank'>Not enough content</a>:" +
@@ -17,7 +17,7 @@ describe( "Assesses presence of text", function() {
 		const mockPaper = new Paper( "Hello. This is a text with over 50 words. We want to see if this assessment is working well." +
 			" What do you think? This is a very interesting paper, with a lot of words. More words are better, don't you think?" +
 			" I am almost done. This is word number fifty. Yay!" );
-		const assessment = textPresence.getResult( mockPaper, Factory.buildMockResearcher( 1 ), i18n );
+		const assessment = new TextPresence().getResult( mockPaper, Factory.buildMockResearcher( 1 ), i18n );
 
 		expect( assessment.getScore() ).toEqual( 0 );
 		expect( assessment.getText() ).toEqual( "" );
@@ -27,7 +27,7 @@ describe( "Assesses presence of text", function() {
 		const mockPaper = new Paper( "Hello. This is a text with exactly 50 words. We want to see if this assessment is working well." +
 			" What do you think? This is a very interesting paper, with a lot of words. More words are better, don't you think?" +
 			" I am almost done. This is word number fifty. Yay!" );
-		const assessment = textPresence.getResult( mockPaper, Factory.buildMockResearcher( 4 ), i18n );
+		const assessment = new TextPresence().getResult( mockPaper, Factory.buildMockResearcher( 4 ), i18n );
 
 		expect( assessment.getScore() ).toEqual( 0 );
 		expect( assessment.getText() ).toEqual( "" );
